Move version to constants package
